v0.04
File simplex.py deleted. Also, 'simplex_maa.ipynb' has been modified, and almost everything before the stepsize calculation has been formatted to be a function.
A main function is included at the end of this file and is to be as close as possible to the pseudo-code of the simplex algorithm.